requirements.txt
fastapi[all]
uvicorn
sqlalchemy
python-dotenv
imaplib2
psycopg2-binary
